Adds CoordinatedBulkWriteSimpleTest

+++ scripts/build-sidecar.sh
@@ -20,23 +20,32 @@
 set -ex
 if [ "${SKIP_SIDECAR_BUILD}" == "true" ]; then
   echo "Skipping Sidecar Build because SKIP_SIDECAR_BUILD was set to 'true'"
+  return 0
+fi
+
+SCRIPT_DIR=$( dirname -- "$( readlink -f -- "$0"; )"; )
+SIDECAR_JAR_DIR="$(dirname "${SCRIPT_DIR}/")/dependencies"
+SIDECAR_JAR_DIR=${CASSANDRA_DEP_DIR:-$SIDECAR_JAR_DIR}
+SIDECAR_BUILD_DIR="${SIDECAR_JAR_DIR}/sidecar-build"
+java_ver_output=`"${JAVA:-java}" -version 2>&1`
+jvmver=`echo "$java_ver_output" | grep '[openjdk|java] version' | awk -F'"' 'NR==1 {print $2}' | cut -d\- -f1`
+JVM_VERSION=${jvmver%_*}
+echo $JVM_VERSION
+if [ "${JVM_VERSION}" == "1.8.0" ]; then
+  SIDECAR_BUILD_VERSION="1.0.0-jdk8-analytics"
 else
-  SCRIPT_DIR=$( dirname -- "$( readlink -f -- "$0"; )"; )
+  SIDECAR_BUILD_VERSION="1.0.0-analytics"
+fi
+
+# Build from local sidecar repo. For instance,
+# LOCAL_SIDECAR_REPO=/PATH/TO/cassandra-sidecar ./scripts/build-sidecar.sh
+if [ "x${LOCAL_SIDECAR_REPO}" != "x" ]; then
+  cd ${LOCAL_SIDECAR_REPO}
+else
   SIDECAR_REPO="${SIDECAR_REPO:-https://github.com/apache/cassandra-sidecar.git}"
   SIDECAR_BRANCH="${SIDECAR_BRANCH:-trunk}"
   SIDECAR_COMMIT="${SIDECAR_COMMIT:-f07e248d0ce8303a06daf93b462190ef7be7304d}"
-  SIDECAR_JAR_DIR="$(dirname "${SCRIPT_DIR}/")/dependencies"
-  SIDECAR_JAR_DIR=${CASSANDRA_DEP_DIR:-$SIDECAR_JAR_DIR}
-  SIDECAR_BUILD_DIR="${SIDECAR_JAR_DIR}/sidecar-build"
-  java_ver_output=`"${JAVA:-java}" -version 2>&1`
-  jvmver=`echo "$java_ver_output" | grep '[openjdk|java] version' | awk -F'"' 'NR==1 {print $2}' | cut -d\- -f1`
-  JVM_VERSION=${jvmver%_*}
-  echo $JVM_VERSION
-  if [ "${JVM_VERSION}" == "1.8.0" ]; then
-    SIDECAR_BUILD_VERSION="1.0.0-jdk8-analytics"
-  else
-    SIDECAR_BUILD_VERSION="1.0.0-analytics"
-  fi
+
   if [[ "$CLEAN" == "true" ]]; then
     echo "Clean up $SIDECAR_BUILD_DIR and $SIDECAR_JAR_DIR/org/apache/cassandra/(cassandra-)sidecar directories"
     rm -rf "${SIDECAR_BUILD_DIR}"
@@ -76,7 +85,12 @@ else
     git checkout "${SIDECAR_COMMIT}"
   fi
   git clean -fd
-  ./gradlew -Pversion=${SIDECAR_BUILD_VERSION} -Dmaven.repo.local=${SIDECAR_JAR_DIR} publishToMavenLocal
-  # Delete sidecar source after publishing to avoid confusing IDE
+fi
+
+# publish to maven local at {SIDECAR_JAR_DIR}
+./gradlew -Pversion=${SIDECAR_BUILD_VERSION} -Dmaven.repo.local=${SIDECAR_JAR_DIR} publishToMavenLocal
+
+if [ "x${LOCAL_SIDECAR_REPO}" == "x" ]; then
+  # Delete the cloned sidecar source after publishing to avoid confusing IDE
   rm -rf "${SIDECAR_BUILD_DIR}"
 fi
